remove stale jobs after a reboot
From bug report:

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=885891

This patch assumes that /run/systemd have a date very near the boot time
of the machine.

Needed to implement the variable atjobdir.

"ExecStartPre=-find @atjobdir@ -type f ... -delete" does not have a
slash on the end of the path on purpose.  If atjobdir is miss
calculated and ends empty the find will fail and do not delete files
from the root.